refactor: move CombinedPrintQueueSink to Application layer
Relocates CombinedPrintQueueSink from the API project (wrong layer)
to Anela.Heblo.Application/Features/ExpeditionList/Services/, matching
the placement of FileSystemPrintQueueSink and restoring Clean Architecture
compliance — no IPrintQueueSink implementations remain in the HTTP shell.

- New namespace: Anela.Heblo.Application.Features.ExpeditionList.Services
- Removed dead `using Anela.Heblo.API.Features.ExpeditionList` from
  ServiceCollectionExtensions.cs (covered by existing Services using)
- Added InternalsVisibleTo("Anela.Heblo.API") to Application assembly
  (AssemblyInfo.cs + .csproj) so internal sealed class is reachable from
  the DI registration call site without widening visibility
- Updated test using statement; no test logic changed
