+++ backend/src/Anela.Heblo.API/Extensions/ServiceCollectionExtensions.cs
@@ -21,7 +21,6 @@ using Anela.Heblo.Adapters.Azure.Features.ExpeditionList;
 using Anela.Heblo.Adapters.Azure;
 using Anela.Heblo.Adapters.Cups;
 using Anela.Heblo.Adapters.Cups.Features.ExpeditionList;
-using Anela.Heblo.API.Features.ExpeditionList;
 using Anela.Heblo.API.PDFPrints;
 using Anela.Heblo.Application.Features.BackgroundJobs.Services;
 using Anela.Heblo.Application.Features.ExpeditionList.Services;
