+++ backend/test/Anela.Heblo.Tests/Features/ExpeditionList/CombinedPrintQueueSinkTests.cs
@@ -1,4 +1,4 @@
-using Anela.Heblo.API.Features.ExpeditionList;
+using Anela.Heblo.Application.Features.ExpeditionList.Services;
 using Anela.Heblo.Application.Shared.Printing;
 using Moq;
